+++ src/main/java/org/truffleruby/core/array/ArrayGuards.java
@@ -18,7 +18,8 @@ public class ArrayGuards {
     // Partial Escape Analysis only considers < 32 elements
     public static final int ARRAY_MAX_EXPLODE_SIZE = 16;
 
-    // Enough to handle all array strategies (all types, plus null and Object[] without longs).
+    // Enough to handle all array strategies (all types, plus null and Object[] without longs plus
+    // delegated storage variants).
     public static final int STORAGE_STRATEGIES = 11;
 
     // Enough to handle all combinations of two strategies.
