+++ src/main/java/org/truffleruby/core/array/DelegatedArrayStorage.java
@@ -1,9 +1,18 @@
+/*
+ * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved. This
+ * code is released under a tri EPL/GPL/LGPL license. You can use it,
+ * redistribute it and/or modify it under the terms of the:
+ *
+ * Eclipse Public License version 1.0
+ * GNU General Public License version 2
+ * GNU Lesser General Public License version 2.1
+ */
 package org.truffleruby.core.array;
 
 public class DelegatedArrayStorage {
-    public Object storage;
-    public int offset;
-    public int length;
+    public final Object storage;
+    public final int offset;
+    public final int length;
 
     public DelegatedArrayStorage(Object storage, int offset, int length) {
         assert offset >= 0;
